Add --prefill-cache, --models flags and 9-model pool support
- Add --prefill-cache to train and evaluate CLI to skip encoder extraction
  when pre-extracted features are available
- Add --models subset filter to train, evaluate, serve, and all inference
  paths (HTTP route, LiteLLM strategy, chat, completions endpoints)
- Fix bfloat16 compatibility in transforms.raw_hidden and evaluate
- Replace manual enabled_models filtering in chat.py with route(models=...)
- Add v1-9models.yaml config (9 models from research parquet pool)
- Add data prep script (prepare_research_data.py) for research asset conversion
- Add reproduce_v1_checkpoint.sh and train_subset.sh scripts
- Add test_model_subset.py and extend adapter tests for models filter
- 101 tests pass, 4 skipped

Made-with: Cursor

+++ src/model_router_toolkit/prefill/transforms.py
@@ -20,7 +20,7 @@ def raw_hidden(result: PrefillResult, layer: int, mode: str) -> np.ndarray:
     """Pull the right hidden state array from a PrefillResult as numpy."""
     tensor = result.hidden_mean[layer] if mode == "mean" else result.hidden_last[layer]
     if isinstance(tensor, torch.Tensor):
-        return tensor.numpy()
+        return tensor.float().numpy()
     return tensor
